+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 * #2: `append($item)` methods return `$item` for chains
 * #3: the directive name can be NULL (it leads to disable directive)
+* #4: don't show the directive comment if the directive is disabled
 
 ## 0.1.0 (11 June 2021)